Written by: Sebastian Reitenbach <sebastia@openbsd.org>
Import the "traditional" GNUstep Objective-C runtime:

The GNUstep Objective-C runtime is designed as a drop-in replacement for the
GCC runtime.

This version of the runtime is needed to get GNUstep ports working fine on sparc.
Not yet hooked up to the build. To make use of it, it needs an update for gnustep-make, and a bump on all dependent ports.
Waiting for gnustep-base update, which should arrive soon, to do all those things at the same time.

OK landry@