Written by: Remi Pointel <rpointel@openbsd.org>
Import cython, from James Turner, with few tweaks by me and landry@.
ok landry@, benoit@. 
Cython is a language that makes writing C extensions for the Python
language as easy as Python itself. Cython is based on the well-known
Pyrex, but supports more cutting edge functionality and optimizations.
The Cython language is very close to the Python language, but Cython
additionally supports calling C functions and declaring C types on
variables and class attributes. This allows the compiler to generate
very efficient C code from Cython code.
This makes Cython a language for wrapping external C libraries, and for
fast C modules that speed up the execution of Python code.