Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import gupnp 0.16.1

GUPnP is an elegant, object-oriented open source framework for creating
UPnP devices and control points, written in C using GObject and libsoup.
The GUPnP API is intended to be easy to use, efficient and flexible. It
provides the same set of features as libupnp, but shields the developer
from most of UPnP's internals.

ok aja@