Written by: Matthias Kilian <kili@openbsd.org>
This filter is designed to replace both the pstoraster and the
pdftoraster filter.
It honours the *LandscapeOrientation: Plus90 | Minus90 | Any PPD
directive (Any interpreted as Plus90) and does the necessary
transformation by ordinary PostScript means, thus bypassing the
special CTM handling by Ghostscript's cups device (which is equivalent
to Minus90).
Currently, only the Landscape/Portrait orientation is handled,
forcing the output to the orientation specified by the PPD default
page size. Any fit-to-page scaling is not yet implemented.

ok aja@