Written by: Matthias Kilian <kili@openbsd.org>
Update to lilypond-2.14.1.

This still uses the downloaded documentation instead of building
it, because it's no fun do build the documentation on low-memory
and/or slow systems like arm (it ran for a couple of days on my
armish, and then it failed with a segfault, which i then couldn't
reproduce for the specific file the segfaults happend on).

It also doesn't depend on the texinfo-4.13 port I sent to ports@
(but not yet committed) recently.

Running it on some of my (rather oldish, i.e. written for lilypond-2.10
and older) files produces mixed results -- simple stuff still works,
some stuff reports failure (yet creates correct output, which is
strange).  Complicated music like my (never finished) work on the
Busoni transcription of BWV-1004/Partita II doesn't work at all
(but iirc, it didn't with lilypond-2.12, either).

make regress currently fails with this version, because it would
need the texinfo port and a new test-baseline, which I didn't yet
create.

Some work to build the documentation for real (possibly shielded by a
pseudo flavor) will follow.