Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import orca-3.1.3.
(devel version, but it's the only one not depending on bonobo and other
deprecated libs)

Orca is a free, open source, flexible, and extensible screen reader
that provides access to the graphical desktop via user-customizable
combinations of speech, braille, and/or magnification.
Orca works with applications and toolkits that support the assistive
technology service provider interface (AT-SPI).

ok jasper@