Written by: Sebastian Reitenbach <sebastia@openbsd.org>
Update gnustep-make to 2.6.1. While there, do not disable native-objective-c
exceptions. They are not used by any port, but maybe someone wants to do it
in the future, or a port may need it.
Further, zap the stripping of binaries that gets propagated into the gnusteps
config.make to be more inline with upstream. This is now properly handled via
the gnustep.port.mk.

OK landry@