Written by: Jeremy Evans <jeremy@openbsd.org>
Update to 1.6.3.  Make bin/jruby a compiled file instead of a bash
script, so it can now be used in shebang lines.  Remove the
building of nailgun, since nailgun support is now part of the
jruby binary.

This requires patching the included jruby-launcher gem to make a
similar change we used to make to the jruby.bash script, to set the
JAVA_HOME environment so things work correctly.

Also, create the SITEARCHDIR directory by default.

OK jcs@