Written by: Stuart Henderson <sthen@openbsd.org>
Permit utimensat(2) and futimens(2), as discussed with matthew@.
Fixes ruby 1.9 build with systrace now that we have these system calls.
The other new *at system calls need to have path restrictions and will
need further work so are still prevented for now (in those cases, the
supplied paths are *relative to a certain FD*, so we can't simply
examine supplied paths).