Written by: Matthias Kilian <kili@openbsd.org>
Switch back to compiling Setup.l?hs and running the resulting program
instead of using runghc (the interpreter).

This is a *workaround* for the following two symptoms:

- Setup.hs: waitForProcess: resource exhausted (Resource temporarily
  unavailable)

- Setup.hs: fd:15: hGetContents: illegal operation (Inappropriate
  ioctl for device)

Note that this is not a *fix*. I'm sure there's some misbehaving
signal handler, either in runghc/ghci or in the ghc runtime library.

(No, it's *not* rts/posix/Signals.c:generic_handler(); I patched
it to preserve errno, but it didn't help)

Until I find the bug, it's better to let bulk builds pass without
(or with less) errors.

ok (with the workaround) espie@