Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import scala 2.9.0.1

Scala is a general purpose programming language designed to express
common programming patterns in a concise, elegant, and type-safe way. It
smoothly integrates features of object-oriented and functional
languages, enabling Java and other programmers to be more productive.
Code sizes are typically reduced by a factor of two to three when
compared to an equivalent Java application.

ok sthen@