Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import msp430mcu-20110613, ok aja@

msp430mcu provides C headers and binutils linker scripts for each of the
over 300 microcontrollers in the MSP430 family.

This is part of the 'uniarch' effort to 'rejoin' mspgcc4 and mspgcc, and
lowering the workload for adding new chips by taking it mostly out of binutils/
gcc and putting it all in one single package. For more information please
refer to http://sourceforge.net/apps/mediawiki/mspgcc/index.php?title=Devel:Uniarch for a full list of changes/rationale behind this.

The rest of the msp430 toolchain will soon be updated to the uniarch versions.