Written by: Marc Espie <espie@openbsd.org>
show FLAVOR to dump-vars.
This is occasionally useful for pseudo-flavors: these do not get encoded
in the pkgpath, so taking (for instance) sqlports, this generates lines
which are later impossible to exploit based only on the fullpkgpath, as
opposed to fullpkgpath,flavor  (which might contain the flavor twice, but
this is not an issue).