Written by: Sebastian Reitenbach <sebastia@openbsd.org>
update to 1.2.3, only minor changes, from the release announcement:


As the version number shows there are no big changes. The most
interesting are the two median filters for tracks. The first one is
optional and will smooth the elevation to give you better values for the
ascend and descend. The second one one is used for the speed over
distance plot. As median filters are no linear filters the result is
much better for the non linear character of an optimized track.

With this release Albrecht's optimizations for the map edit/create
dialogs are becoming official. Something all netbook users will appreciate.

OK jasper@