Written by: Jasper Lievisse Adriaanse <jasper@openbsd.org>
import arduino 22

Arduino is an open-source electronics prototyping platform based on
flexible, easy-to-use hardware and software. It's intended for artists,
designers, hobbyists, and anyone interested in creating interactive
objects or environments. 
[..]
This package contains the libraries, header files and tools to develop
Arduino sketches. It does not include the Arduino IDE frontend.

based on work from and ok ckuethe@