Written by: Landry Breuil <landry@openbsd.org>
Import devel/p5-File-RsyncP 0.70, needed by BackupPC.

File::RsyncP is a perl implementation of an Rsync client.  It is
compatible with Rsync 2.5.5 (protocol version 26).  It can send
or receive files, either by running rsync on the remote machine,
or connecting to an rsyncd deamon on the remote machine.
ok sthen@