Written by: Landry Breuil <landry@openbsd.org>
Import BackupPC 3.2.1.

BackupPC is a high-performance, enterprise-grade system for backing
up Unix, WinXX, and MacOS PCs and laptops to a server's disk.
BackupPC is highly configurable and easy to install and maintain.
It is written in Perl and extracts backup data via rsync, or tar
over ssh/rsh/nfs.

With tweaks from and ok aja@