Written by: Giovanni Bechis <giovanni@openbsd.org>
Import p5-Mail-SPF, a perl oop implementation of Sender Policy Framework
ok sthen@

Mail::SPF is an object-oriented implementation of Sender Policy
Framework (SPF).  This class collection aims to fully conform to
the SPF specification (RFC 4408) so as to serve both as a production
quality SPF implementation and as a reference for other developers
of SPF implementations.