Written by: Stuart Henderson <sthen@openbsd.org>
amsn requires internal headers from libpng; currently this requires a manual
change for every update to png to pass in the correct directory version to
configure. Change this to create a symlink to the dir in pre-configure, and
pass the path to the symlink to configure instead, so png updates don't need
to touch this. Discussed with naddy@.