Written by: Martin Pieuchot <mpi@openbsd.org>
Import py-ipy 0.75

IPy is a Python module for handling IPv4 and IPv6 Addresses and 
Networks in a fashion similar to perl's Net::IP and friends.

The IP class allows a comfortable parsing and handling for most 
notations in use for IPv4 and IPv6 Addresses and Networks. 

It can detect about a dozen different ways of expressing IP addresses
and networks, parse them and distinguish between IPv4 and IPv6
addresses.

ok rpointel@