Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Import rubber-1.1.

Rubber is a program whose purpose is to handle all tasks related to the
compilation of LaTeX documents. This includes compiling the document
itself, of course, enough times so that all references are defined, and
running BibTeX to manage bibliographic references. Automatic execution
of dvips to produce PostScript documents is also included, as well as
usage of pdfLaTeX to produce PDF documents.

ok jasper@