Written by: Matthias Kilian <kili@openbsd.org>
Use the new bootstrappers (thanks to sthen@ for doing the i386
bootstrapper).

Bump, because some of the `library hashes' changed.

Unfortunately, this also affects a couple of other haskell ports,
which get different `library hashes' now and will be reported as
broken by ghc-pkg check.

People using snapshot packages should get seamless updates when the
next set of snapshots is available. People building from ports and
*not* using dpb should use the out-of-date-script to get a subdirlist
of haskell ports which should be rebuilt.

I could also bump the affected ports, but I'm not sure wether it's
worth the trouble.