Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Sync dependencies after move from libusb to libusb-compat and fix a
couple of fallouts. Note that these ports are only the ones that used to
have a direct dependency on devel/libusb so there may be some other
hidden ports that may break because of the switch. If that's the case,
no need to start ranting all over but instead tell me which one(s)
break. Thanks.

Note: sysutils/nut hasn't been fixed yet but will be today.