Written by: Jeremy Evans <jeremy@openbsd.org>
Make regress tests work.

Set JAVA_HOME and PATH and disable forking so java doesn't crash when
running the test and spec tasks.  Don't install the jruby-launcher gem
when running the test and spec tasks, as it fails.

Regress depend on git when running the specs, as it needs to download
current rubyspecs for the spec task.

Delete the jruby-complete.jar file and just use the jruby.jar file
when running the specs, as otherwise you get additional test failures
and warnings.

Switch to SUBST_CMD while here now that has been fixed. Also, bump
the memory limits by 50M as 384M doesn't seem to be enough in all
cases.