Written by: Jeremy Evans <jeremy@openbsd.org>
Split rspec port into subdirs.  Version 1 is still a single package,
but version 2 has 4 packages (core, expectations, mocks, and rspec).
Stick version 1 in a subdir named 1, and the version 2 packages
each in their own subdir.

Upgrade version 1 to 1.3.2, the latest version, as some ports depend
on >=1.3.0,<2.0.  Because rubygems does not correctly handle the
case where two versions of the same library install different
binaries, manually hack the version 1 spec binary to work.

Both version 1 and version 2 ship with the autospec binary, so comment
it out from version 1 so the versions don't conflict.

This requires changes to dependent ports, which will be committed
shortly.