Written by: Matthias Kilian <kili@openbsd.org>
Guppy-PE is a library and programming environment for Python,
currently providing in particular the Heapy subsystem, which supports
object and heap memory sizing, profiling and debugging. It also
includes a prototypical specification language, the Guppy Specification
Language (GSL), which can be used to formally specify aspects of
Python programs and generate tests and documentation from a common
source.

ok rpointel@