Written by: Jeremy Evans <jeremy@openbsd.org>
Update to 3.0.11.  Support by the standalone version of passenger by
splitting into a multipackage, with a subpackage for the standalone
version (which embeds a version of nginx).  The standalone version
operates much like other ruby webservers, serving a single ruby/rack
application.

Switch to using the gem version of passenger.  Because the gem
installs into a versioned directory, setup symlinks to the
versioned directory so that nginx configuration files don't need
to be modified when the version is updated.