Written by: Matthias Kilian <kili@openbsd.org>
Update to haskell-plattform-2011.4.0.0.

This does not (yet) contain the required update of
xhtml, because an older xhtml library is still part
of the ghc package. This weil be fixed next year.

Please note that you have to pkg_delete this meta-package
before trying to build the new version and/or the updated
hs- libraries.

ok jasper@