Written by: Matthias Kilian <kili@openbsd.org>
This package contains code for generating high quality random numbers
that follow either a uniform or normal distribution.  The generated
numbers are suitable for use in statistical applications.

The uniform PRNG uses Marsaglia's MWC256 (also known as MWC8222)
multiply-with-carry generator, which has a period of 2^8222 and
fares well in tests of randomness.  It is also extremely fast,
between 2 and 3 times faster than the Mersenne Twister.

Compared to the mersenne-random package, this package has a more
convenient API, is faster, and supports more statistical distributions.

ok sthen@