Written by: Marc Espie <espie@openbsd.org>
switch to a new way of bundling bootstrap: this bundles the official
libc.so from OpenBSD 5.0, and tweaks the ada scripts to LD_LIBRARY_PATH
the bundled library.

That way, we should be able to get away by providing a new bootstrap ways
less often (preferably every release, but things won't break if we do it
even LESS often).