Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Compile the gobject parts with -O0 on 64bit arches. There is a nasty
SEGFAULT when using g_closure() on at least amd64; unclear whether this
is a gcc/binutils/glib... bug yet but we want to be on the safe side for
the release.
Hopefully someone will be able to figure this out, I haven't...

Thanks to Stuart for teaching me PROPERTIES.
ok sthen@ naddy@