Written by: Matthias Kilian <kili@openbsd.org>
Something is wrong with the threaded rts and rthreads, so build ghc
itself using the non-threaded rts. This doesn't fix the real problem,
but it makes it possible to build hs-* ports without filling up
dpb's build slots with stuck ghc processes.

ok naddy@ espie@

Note: I'll re-run the testsuite with this diff and hope to get some
reasonable results that may help us to identify the real problem.