Written by: Antoine Jacoutot <ajacoutot@openbsd.org>
Major update to glib2-2.31.20.
Note that this is a devel version but the new stable is expected within a
week or so. Committing now so that I have time to deal with the breakage
and can start the upcoming gnome3 update.

Please contact me _directly_ if you find any issue with it.