Written by: Sebastian Reitenbach <sebastia@openbsd.org>
updates HighlighterKit from 0.1.1 to 0.1.2. Getting rid of a bunch of patches, and now the new version allows to use themes for the highlighting.

Added a LIBDIR variable to the SUBST_VARS to not have a large bunch of changes in the PLIST when the SHLIB_VERSION changes the next time.

OK aja@